+++ gui/src/app/projects/page.tsx
@@ -11,14 +11,13 @@ import CreateOrEditProjectBody from '@/components/HomeComponents/CreateOrEditPro
 import CustomLoaders from '@/components/CustomLoaders/CustomLoaders';
 import { SkeletonTypes } from '@/app/constants/SkeletonConstants';
 import CustomImage from '@/components/ImageComponents/CustomImage';
-import CustomModal from '@/components/CustomModal/CustomModal';
+import GithubStarModal from '@/components/HomeComponents/GithubStarModal';
 
 export default function Projects() {
   const [openNewProjectModal, setOpenNewProjectModal] = useState<
     boolean | null
   >(false);
   const [projectsList, setProjectsList] = useState<ProjectTypes[] | null>(null);
-  const [showGithubStar, setShowGithubStar] = useState<boolean | null>(true);
   const router = useRouter();
 
   const handleProjectClick = async (project: ProjectTypes) => {
@@ -26,42 +25,8 @@ export default function Projects() {
     router.push(`/board`);
   };
 
-  const handleGithubStarClick = () => {
-    window.open('https://github.com/TransformerOptimus/SuperCoder', '_blank');
-    setModalInteraction();
-  };
-
-  const handleGithubStarClose = () => {
-    setModalInteraction();
-  };
-
-  const setModalInteraction = () => {
-    const fourWeeksFromNow = new Date();
-    fourWeeksFromNow.setDate(fourWeeksFromNow.getDate() + 28);
-    localStorage.setItem(
-      'githubStarHiddenTill',
-      fourWeeksFromNow.toISOString(),
-    );
-    setShowGithubStar(false);
-  };
-
-  const checkModalInteraction = () => {
-    const closedUntil = localStorage.getItem('githubStarHiddenTill');
-    if (closedUntil) {
-      const closedUntilDate = new Date(closedUntil);
-      return closedUntilDate > new Date();
-    }
-    return false;
-  };
-
   useEffect(() => {
     toGetAllProjects(setProjectsList).then().catch();
-  }, []);
-
-  useEffect(() => {
-    if (checkModalInteraction()) {
-      setShowGithubStar(false);
-    }
   }, []);
 
   return (
@@ -76,39 +41,7 @@ export default function Projects() {
         projectsList={projectsList}
       />
 
-      <CustomModal
-        isOpen={showGithubStar}
-        onClose={handleGithubStarClose}
-        width={'32vw'}
-      >
-        <CustomModal.Body padding={'24px'}>
-          <div
-            className={
-              'proxima_nova flex flex-col items-center gap-3 text-white'
-            }
-          >
-            <div className={'text-base font-medium'}>
-              Support the project by leaving a star on GitHub repository
-            </div>
-
-            <Button
-              className={'secondary_medium mt-3 w-fit'}
-              onClick={handleGithubStarClick}
-            >
-              Leave a ⭐ star on GitHub
-            </Button>
-
-            <div
-              className={
-                'secondary_color cursor-pointer text-xs font-normal opacity-80'
-              }
-              onClick={handleGithubStarClose}
-            >
-              I’ll do it later
-            </div>
-          </div>
-        </CustomModal.Body>
-      </CustomModal>
+      <GithubStarModal />
 
       {projectsList ? (
         <div className={'flex flex-col gap-6'}>
